deepface-autoencoder/requirements.txt
torch
torchvision
pillow
matplotlib
pytorch-msssim
